7th task and readme update

+++ 0x06-pointers_arrays_strings/6-cap_string.c
@@ -16,7 +16,7 @@ char *cap_string(char *str)
 	{
 		int j = 0;
 
-		for (; j < sizeof(separators); j++)
+		for (; j < (int)sizeof(separators); j++)
 		{
 			if (str[i] >= 97 && str[i] <= 122)
 			{
